Port to SDL2
Also add a -DSDL flag to allow #ifdef SDL sections.
SDL2 is found via pkg-config.
I didn't manage to get the linking right using CMake.
Building with `make VERBOSE=1` and appending a -lSDL2
to the command line creates a working binary.
Keeping -lSDL2 in the same place and changing it to
-Bdynamic -lSDL2 -Bstatic also works.